chore: moves RequestHeaders into propagation folder, also adds typing to the default parsers.
